refactor: standardize file naming conventions for components
- Renamed UI component files to use PascalCase following React conventions
- Fixed imports to reference the new PascalCase filenames
- Renamed utility file from kebab-case to camelCase
