+++ src/components/ui/Button.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Slot } from './simple-slot';
+import { Slot } from './SimpleSlot';
 import type { VariantProps } from 'class-variance-authority';
 
 import { cn } from '@/lib/utils';
@@ -23,6 +23,7 @@ const Button = React.forwardRef<HTMLButtonElement, ButtonProps>(
     );
   }
 );
+
 Button.displayName = 'Button';
 
-export { Button };
+export { Button, buttonVariants };
