+++ src/components/ui/Input.tsx
@@ -18,6 +18,7 @@ const Input = React.forwardRef<
     />
   );
 });
+
 Input.displayName = 'Input';
 
 export { Input };
