+++ src/components/ui/questionCounter/SmallCircularQuestionCounter.tsx
