+++ src/features/wizard/components/steps/PrivacyStep.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import StepContainer from '../StepContainer';
-import { Button } from '@/components/ui/button';
+import { Button } from '@/components/ui/Button';
 import { MailX, MailPlus } from 'lucide-react';
 import { useEntries } from '@/features/statements/hooks/useEntries';
